requirements.txt
openpyxl>=3.1.2
Faker>=25.0.0
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
